[TAN-8637] Move the events settings panel to its own file
