+++ front/app/components/ProjectPageBuilder/Widgets/Events/index.tsx
@@ -1,16 +1,13 @@
 import React, { useEffect } from 'react';
 
-import { Box, Text } from '@citizenlab/cl2-component-library';
+import { Box } from '@citizenlab/cl2-component-library';
 import { UserComponent } from '@craftjs/core';
 
 import { maxPageWidth } from 'containers/ProjectsShowPage/styles';
 
 import useCraftComponentDefaultPadding from 'components/admin/ContentBuilder/useCraftComponentDefaultPadding';
-import useWidgetProjectId from 'components/admin/ContentBuilder/useWidgetProjectId';
 import SharedEventsWidget from 'components/admin/ContentBuilder/Widgets/Events';
 
-import { FormattedMessage } from 'utils/cl-intl';
-import Link from 'utils/cl-router/Link';
 import { useLocation, useParams } from 'utils/router';
 import { scrollToElement } from 'utils/scroll';
 
@@ -19,8 +16,9 @@ import messages from '../messages';
 import SectionBackground, {
   SectionBackgroundChoice,
 } from '../SectionBackground';
-import SectionBackgroundSetting from '../SectionBackgroundSetting';
 import useIsPageBodyChild from '../useIsPageBodyChild';
+
+import EventsSettings from './Settings';
 
 export const EVENTS_WIDGET_ANCHOR_ID = 'e2e-project-page-events';
 
@@ -64,34 +62,6 @@ const EventsWidget: UserComponent<Props> = ({ sectionBackground }) => {
         </EditModeHeightCap>
       )}
     />
-  );
-};
-
-const EventsSettings = () => {
-  const projectId = useWidgetProjectId();
-
-  return (
-    <Box my="20px">
-      <SectionBackgroundSetting defaultValue="white" />
-      <Text color="textSecondary" fontSize="s">
-        <FormattedMessage
-          {...messages.eventsManagedNote}
-          values={{
-            eventsLink: projectId ? (
-              <Link
-                to="/admin/projects/$projectId/events"
-                params={{ projectId }}
-                target="_blank"
-              >
-                <FormattedMessage {...messages.eventsLinkText} />
-              </Link>
-            ) : (
-              <FormattedMessage {...messages.eventsLinkText} />
-            ),
-          }}
-        />
-      </Text>
-    </Box>
   );
 };
